Ajuste nos calculos, e inicio da implementação de autocomplete , falta testes unitarios do select autocomplete e finalizar a front.
