+++ apps/finance/src/app/bills/components/expenses/persist/Persist.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
+import { useRouter } from 'next/navigation';
 
 import { EMonth, MONTHS, ValidatorParams, convertToNumber } from '@repo/services';
 
@@ -37,6 +38,7 @@ export default function Persist({
     onSubmit,
 }: PersistProps) {
     const isMounted = useRef(false);
+    const router = useRouter();
 
     const [inputGroups, setInputGroups] = useState<Array<InputGroup>>([]);
     const [inputs, setInputs] = useState<Array<InputGroupItem>>([]);
@@ -249,6 +251,23 @@ export default function Persist({
         return input.validator(params);
     }
 
+    const fallbackAction = (name: string) => {
+        switch (name) {
+            case 'supplier':
+                router.push('/suppliers');
+                break;
+            default:
+                break;
+        }
+    }
+
+    const fallbackLabel = (name?: string) => {
+        if(name === 'supplier') {
+            return 'Add Supplier';
+        }
+        return;
+    }
+
     useEffect(() => {
         if(!isMounted.current) {
             isMounted.current = true;
@@ -275,6 +294,8 @@ export default function Persist({
                                     onInput={handleOnInput}
                                     className={input.className ?? 'persist__row--item'}
                                     validator={(params) => handleValidator(input, params)}
+                                    fallbackLabel={fallbackLabel(input?.name)}
+                                    fallbackAction={fallbackAction}
                                 />
                             )}
                             {show && input.label === 'Paid' && (
