+++ packages/business/src/api/nest/finance/bill/index.ts
@@ -8,7 +8,6 @@ export * from './enum';
 export {
     type IExpense,
     type IExpenseBase,
-    type IExpenseMonthsWithPaid,
     type ICreateExpenseParams,
     type IUpdateExpenseParams,
     EExpenseType
