+++ packages/business/src/api/nest/finance/month/types.ts
@@ -30,3 +30,30 @@ export type IPersistMonthParams = Partial<Omit<IMonth, 'value'>> & {
     value: number;
     month?: EMonth;
 };
+
+export type IMonthsObject = {
+    january: number;
+    january_paid: boolean;
+    february: number;
+    february_paid: boolean;
+    march: number;
+    march_paid: boolean;
+    april: number;
+    april_paid: boolean;
+    may: number;
+    may_paid: boolean;
+    june: number;
+    june_paid: boolean;
+    july: number;
+    july_paid: boolean;
+    august: number;
+    august_paid: boolean;
+    september: number;
+    september_paid: boolean;
+    october: number;
+    october_paid: boolean;
+    november: number;
+    november_paid: boolean;
+    december: number;
+    december_paid: boolean;
+}
