+++ packages/business/src/finance/expense/business/business.ts
@@ -1,47 +1,28 @@
-import {
-    cleanTextByListText,
-    type CycleOfMonths,
-    DEFAULT_TABLES_PARAMS,
-    getCurrentMonth,
-    getCurrentMonthNumber,
-    getMonthByIndex,
-    getMonthIndex,
-    isMonthValid,
-    MONTHS,
-    snakeCaseToNormal,
-    type TablesParams,
-    TMonth
-} from '@repo/services';
+import { getCurrentMonth, getMonthByIndex, getMonthIndex, isMonthValid, MONTHS, TMonth } from '@repo/services';
 
-import { EBillType, EExpenseType } from '../../../api';
+import { Month, MonthBusiness, MonthsCalculated } from '../../month';
 
-import { ExpenseEntity, ExpenseMonthsWithPaid, ExpenseWithMonthsAndPaid, InitializedExpense } from '../types';
+import { ExpenseEntity, ExpenseWithMonthsAndPaid, InitializedExpense } from '../types';
 import type Expense from '../expense';
 
-import type {
-    AccumulateGroupTables,
-    AccumulateGroupTablesParams,
-    AllExpensesCalculated,
-    BuildCreditCardBodyData,
-    BuildCreditCardBodyDataParams,
-    BuildDetailData,
-    BuildDetailDataParams,
-    BuildGroupTable,
-    BuildGroupTableParams,
-    DataAccumulator,
-    GenerateCreditCardTable,
-    GenerateCreditCardTableParams,
-    GenerateDetailsTable,
-    GenerateDetailsTableParams,
-    ParseToDetailsTable,
-    ParseToDetailsTableParams,
-} from './types';
+import { SpreadsheetBusiness } from './spreadsheet';
 
-import type { Bill } from '../../bill';
 
 export default class ExpenseBusiness {
+    private readonly spreadsheetBusiness: SpreadsheetBusiness
+    private readonly monthsBusiness: MonthBusiness;
+
+    constructor() {
+        this.spreadsheetBusiness = new SpreadsheetBusiness();
+        this.monthsBusiness = new MonthBusiness();
+    }
+
+    public get spreadsheet(): SpreadsheetBusiness {
+        return this.spreadsheetBusiness;
+    }
+
     public initialize(expense: Expense, month?: ExpenseEntity['month']): InitializedExpense {
-        return expense.type === EExpenseType.FIXED
+        return expense.type === 'FIXED'
             ? this.handleFixedExpense(expense)
             : this.handleVariableExpense(expense, month);
     }
@@ -53,7 +34,7 @@ export default class ExpenseBusiness {
         months.forEach(month => {
             const existingCurrentMonth = existingExpense.months.find(m => m.label === month);
             const currentMonth = expense.months.find(m => m.label === month);
-            if(existingCurrentMonth && currentMonth) {
+            if (existingCurrentMonth && currentMonth) {
                 existingCurrentMonth.value += currentMonth.value;
                 existingCurrentMonth.paid = currentMonth.paid;
                 const existingCurrentMonthIndex = existingExpense.months.findIndex(m => m.label === month);
@@ -92,11 +73,16 @@ export default class ExpenseBusiness {
             monthsForNextYear,
             expenseForNextYear: undefined,
             monthsForCurrentYear,
-            expenseForCurrentYear: { ...expense, instalment_number: monthsForCurrentYear.length  },
+            expenseForCurrentYear: { ...expense, instalment_number: monthsForCurrentYear.length },
         };
 
         if (result.requiresNewBill) {
-            result.expenseForNextYear = { ...expense, id: '', year: result.nextYear, instalment_number: monthsForNextYear.length };
+            result.expenseForNextYear = {
+                ...expense,
+                id: '',
+                year: result.nextYear,
+                instalment_number: monthsForNextYear.length
+            };
         }
         return result;
     }
@@ -119,47 +105,33 @@ export default class ExpenseBusiness {
         return { monthsForCurrentYear, monthsForNextYear };
     }
 
-    public calculate(expense: Expense): Expense {
+    private calculateChildren(expenses: Array<Expense>, month: Month): Month {
+        return expenses.reduce((acc, expense) => this.monthsBusiness.calculateByMonth(acc, expense?.months), month);
+    }
+
+    private calculateParent(expense: Expense): Expense {
+        if (!expense?.children || expense.children.length === 0) {
+            return expense;
+        }
         const builtExpense = { ...expense };
-        const months = expense.months;
-        builtExpense.paid = months.every(month => month.paid === true);
-        const total = months.reduce((acc, item) => acc + item.value, 0);
-        const total_paid = months.reduce((acc, item) => acc + (item.paid ? item.value : 0), 0);
-        builtExpense.total = Number(total.toFixed(2));
-        builtExpense.total_paid = Number(total_paid.toFixed(2));
+        builtExpense.children = expense.children.map((child) => this.calculate({ ...child, children: undefined }));
+        builtExpense.months = expense.months.map((month) => this.calculateChildren(builtExpense.children, month));
         return builtExpense;
     }
 
-    public buildTablesParams(expenses: Array<Expense> = [], tableWidth: number, headers: Array<string> = ['month', 'value', 'paid']): TablesParams {
-        const tables: TablesParams['tables'] = [];
-
-        expenses.forEach((expense) => {
-            const monthlyData = expense.months.map((month) => ({
-                month: month.label,
-                value: month.value,
-                paid: month.paid,
-            }));
-            const body = {
-                title: expense?.supplier?.name || 'expense',
-                data: monthlyData
-            };
-            tables.push(body);
-        });
-
-        return {
-            ...DEFAULT_TABLES_PARAMS,
-            tables,
-            headers,
-            tableWidth,
-            tableDataRows: MONTHS.length,
-        };
+    public calculate(expense: Expense): Expense {
+        const expenseToCalculate = expense.children?.length > 0 ? this.calculateParent(expense) : expense;
+        const monthsCalculated = this.monthsBusiness.calculateAll(expenseToCalculate?.months);
+        expenseToCalculate.paid = monthsCalculated.allPaid;
+        expenseToCalculate.total = monthsCalculated.total;
+        expenseToCalculate.total_paid = monthsCalculated.totalPaid;
+        expenseToCalculate.total_pending = monthsCalculated.totalPending;
+        return expenseToCalculate;
     }
 
     public totalByMonth(month: string, expenses: Array<Expense> = []): number {
         return expenses.reduce((sum, expense) => {
-            const code = getCurrentMonthNumber(month)
-            const foundMonth = expense.months?.find(m => m.code === code);
-            return sum + (foundMonth?.value || 0);
+            return sum + this.monthsBusiness.totalByMonth(month, expense?.months);
         }, 0);
     }
 
@@ -170,433 +142,24 @@ export default class ExpenseBusiness {
         return expenses.every(expense => expense && expense.paid === true);
     }
 
-    public parseToDetailsTable({
-                                   bills,
-                                   startRow,
-                                   groupName,
-                                   workSheet
-                               }: ParseToDetailsTableParams): ParseToDetailsTable {
-        const secondaryBillList = bills.filter((bill) => bill.type !== EBillType.CREDIT_CARD);
-        const creditCardBillList = bills.filter((bill) => bill.type === EBillType.CREDIT_CARD);
-        const expensesData: ParseToDetailsTable = [];
-
-        const { data: secondaryTableList, nextRow: secondaryTableListNextRow } = this.generateDetailsTable({
-            bills: secondaryBillList,
-            startRow,
-            workSheet
-        });
-
-        if (secondaryTableListNextRow !== startRow) {
-            expensesData.push(...secondaryTableList);
-        }
-
-        const { data: creditCardTableList, nextRow: creditCardTableListNextRow } = this.generateCreditCardTable({
-            bills: creditCardBillList,
-            groupName,
-            startRow: secondaryTableListNextRow,
-            workSheet
-        });
-
-        if (creditCardTableListNextRow !== secondaryTableListNextRow) {
-            expensesData.push(...creditCardTableList);
-        }
-
-        return expensesData;
-
-    }
-
-    private generateDetailsTable({ bills, startRow, workSheet }: GenerateDetailsTableParams): GenerateDetailsTable {
-        const billTypeMap = new Map(bills.map(bill => [bill.type, bill]));
-        const collectGroupsRecursively = (row: number, acc: DataAccumulator): {
-            data: DataAccumulator;
-            nextRow: number
-        } => {
-            const cell = workSheet.cell(row, 3);
-            const type = cell?.value ? cell?.value?.toString()?.trim() : '';
-
-            if (!type) {
-                return { data: acc, nextRow: row };
-            }
-
-            const bill = billTypeMap.get(type as EBillType);
-
-            if (!bill) {
-                return { data: acc, nextRow: row };
-            }
-
-            const { acc: filledAcc, lastRow } = this.accumulateGroupTables({
-                acc,
-                bill,
-                startRow: row + 1,
-                workSheet,
-            });
-
-            return collectGroupsRecursively(lastRow, filledAcc);
-        };
-
-        const { data, nextRow } = collectGroupsRecursively(startRow, []);
-
-        return { data, nextRow };
-    }
-
-    private accumulateGroupTables({
-                                      acc,
-                                      bill,
-                                      startRow,
-                                      workSheet
-                                  }: AccumulateGroupTablesParams): AccumulateGroupTables {
-        const constructedGroupTable = this.buildGroupTable({
-            row: startRow,
-            bill,
-            workSheet,
-        });
-        const updatedAcc = [...acc, ...constructedGroupTable.data];
-        if (constructedGroupTable.hasNext) {
-            return this.accumulateGroupTables({
-                acc: updatedAcc,
-                bill,
-                startRow: constructedGroupTable.nextRow,
-                workSheet
-            });
-        }
-        return { acc: updatedAcc, lastRow: constructedGroupTable.nextRow };
-    }
-
-
-    private buildGroupTable({ row, bill, workSheet }: BuildGroupTableParams): BuildGroupTable {
-        const tableCell1 = workSheet.cell(row, 3);
-        if (tableCell1.isMerged && tableCell1['_mergeCount'] === 2) {
-            const groupTableData: Array<Record<string, string | number | boolean | object | Bill>> = [];
-            const groupTableRow = row + 2;
-            const groupTable1Data1 = this.buildDetailData({
-                row: groupTableRow,
-                cell: tableCell1,
-                bill,
-                column: 4,
-                workSheet,
-            });
-            if (groupTable1Data1) {
-                groupTableData.push(groupTable1Data1);
-            }
-            const tableCell2 = workSheet.cell(row, 8);
-            const groupTable1Data2 = this.buildDetailData({
-                row: groupTableRow,
-                cell: tableCell2,
-                bill,
-                column: 9,
-                workSheet,
-            });
-            if (groupTable1Data2) {
-                groupTableData.push(groupTable1Data2);
-            }
-
-            const tableCell3 = workSheet.cell(row, 13);
-            const groupTable1Data3 = this.buildDetailData({
-                row: groupTableRow,
-                cell: tableCell3,
-                bill,
-                column: 14,
-                workSheet,
-            });
-            if (groupTable1Data3) {
-                groupTableData.push(groupTable1Data3);
-            }
-            return { data: groupTableData, nextRow: row + 12 + 1 + 1 + 1, hasNext: true };
-        }
-
-        return { data: [], nextRow: row, hasNext: false };
-
-    }
-
-
-    private buildDetailData({
-                                row,
-                                cell,
-                                bill,
-                                column,
-                                workSheet,
-                            }: BuildDetailDataParams): BuildDetailData | undefined {
-        const title = cell.value ? cell.value.toString().trim() : '';
-        if (title === '') {
-            return;
-        }
-
-        const bodyData = MONTHS.reduce((acc, month, index) => {
-            const valueCell = workSheet.cell(row + index, column);
-            const valueText = valueCell.value ? valueCell.value.toString().trim() : '0';
-            const paidCell = workSheet.cell(row + index, column + 1);
-            const paidText = paidCell.value ? paidCell.value.toString().trim() : 'NO';
-            return {
-                ...acc,
-                [month]: Number(valueText),
-                [`${month}_paid`]: paidText === 'YES'
-            };
-        }, {} as CycleOfMonths);
-
-        return {
-            ...bodyData,
-            bill,
-            supplier: title
-        };
-    }
-
-
-    private generateCreditCardTable({
-                                        bills,
-                                        startRow,
-                                        groupName,
-                                        workSheet
-                                    }: GenerateCreditCardTableParams): GenerateCreditCardTable {
-        const regex = /^([A-Z_]+)\(([^)]+)\)$/;
-
-        const readExpensesBlock = (
-            bill: Bill,
-            supplierList: Array<string>,
-            currentRow: number,
-            stopValue: string,
-            parent?: DataAccumulator[number]
-        ) => {
-            const inner = (
-                row: number,
-                acc: DataAccumulator
-            ): { expenses: typeof acc, nextRow: number } => {
-                const cell = workSheet.cell(row, 2);
-                const cellValue = cell.value?.toString().trim() || '';
-
-                if (!cellValue || cellValue === stopValue) {
-                    return { expenses: acc, nextRow: row };
-                }
-
-                const { data: bodyData, supplierList: suppliers } = this.buildCreditCardBodyData({
-                    row,
-                    bill,
-                    column: 2,
-                    isParent: !parent,
-                    groupName,
-                    workSheet,
-                    supplierList
-                });
-
-                if (suppliers?.length) {
-                    supplierList.push(...suppliers);
-                }
-
-                const accNext = bodyData
-                    ? [...acc, { ...bodyData, ...(parent ? { parent } : {}) }]
-                    : acc;
-                return inner(row + 1, accNext);
-            };
-            return inner(currentRow, []);
-        };
-
-        const processParentWithChildren = (
-            data: DataAccumulator,
-            initialRow: number,
-            supplierList: Array<string>,
-            bill: Bill
-        ): number => {
-            const recurse = (row: number): number => {
-                const cell = workSheet.cell(row, 2);
-                const value = cell.value?.toString().trim() || '';
-                if (cell.isMerged && cell['_mergeCount'] > 2) {
-
-                    const parentName = `${groupName} ${value}`;
-                    const parent = data.find((item) => item.name === parentName);
-                    if (parent) {
-                        const { expenses: children, nextRow } = readExpensesBlock(
-                            bill,
-                            supplierList,
-                            row + 2,
-                            '',
-                            parent
-                        );
-                        if (children.length) parent.children = children;
-                        return recurse(nextRow + 1);
-                    }
-                }
-                return row;
-            };
-            return recurse(initialRow);
-        };
-
-        const processBills = (
-            row: number,
-            acc: DataAccumulator
-        ): { allBills: typeof acc, nextRow: number } => {
-            const cell = workSheet.cell(row, 2);
-            const cellValue = cell.value?.toString().trim() || '';
-            const match = cellValue.match(regex);
-
-            if (!match) {
-                return { allBills: acc, nextRow: row };
-            }
-
-
-            const billType = match[1] as EBillType;
-
-            const bankName = match?.[2] || 'Bank';
-
-            const billName = `${groupName} ${snakeCaseToNormal(billType)} ${bankName}`;
-
-            const bill = bills.find((item) => item.name === billName);
-
-            if (!bill) {
-                return processBills(row + 1, acc);
-            }
-
-
-            const supplierList: Array<string> = [];
-            const { expenses: data, nextRow } = readExpensesBlock(
-                bill,
-                supplierList,
-                row + 2,
-                'TOTAL'
-            );
-            const afterParentRow = processParentWithChildren(data, nextRow + 2, supplierList, bill);
-            return processBills(
-                afterParentRow,
-                [...acc, { ...bill, expenses: data }]
-            );
-        };
-        const { allBills, nextRow } = processBills(startRow, []);
-
-        const data: DataAccumulator = [];
-        allBills.forEach((bill) => {
-            if (bill['expenses']) {
-                const items = bill['expenses'];
-                if (Array.isArray(items)) {
-                    data.push(...items);
-                }
-            }
-        });
-
-        return {
-            data,
-            nextRow: nextRow
-        };
-    }
-
-    private buildCreditCardBodyData({
-                                        row,
-                                        bill,
-                                        column,
-                                        isParent = true,
-                                        groupName,
-                                        workSheet,
-                                        supplierList = []
-                                    }: BuildCreditCardBodyDataParams): BuildCreditCardBodyData {
-        const filterTexts: Array<string> = [
-            bill.name,
-            ...(!isParent ? supplierList : [])
-        ];
-
-        const titleCellTable = workSheet.cell(row, column);
-        const titleCellTableValue = titleCellTable.value ? titleCellTable.value.toString().trim() : '';
-
-        const name = `${groupName} ${titleCellTableValue}`;
-
-        const monthsObj = MONTHS.reduce((acc, month) => {
-            acc[month] = 0;
-            acc[`${month}_paid`] = false;
+    public calculateAll(expenses: Array<Expense> = []): MonthsCalculated {
+        return expenses.reduce((acc, expense) => {
+            const expenseCalculated = this.calculate(expense);
+            acc.total = acc.total + expenseCalculated.total;
+            acc.allPaid = acc.allPaid && expenseCalculated.paid;
+            acc.totalPaid = acc.totalPaid + expenseCalculated.total_paid;
+            acc.totalPending = acc.totalPending + expenseCalculated.total_pending;
             return acc;
-        }, {} as CycleOfMonths);
-
-        const supplier = cleanTextByListText(filterTexts, name);
-
-        const normalizedSupplierList = isParent ? [supplier] : [];
-
-        const monthValues = MONTHS.map((month, index) => {
-            const currentColumn = column + 1 + index;
-            const monthCell = workSheet.cell(row, currentColumn);
-            const monthCellValue = monthCell?.value ? monthCell?.value?.toString()?.trim() : '0';
-            const monthCellValueNumber = Number(monthCellValue);
-            return { [month]: Number.isNaN(monthCellValueNumber) ? 0 : monthCellValueNumber };
-        });
-
-        const paidColumn = column + 1 + MONTHS.length;
-        const paidCell = workSheet.cell(row, paidColumn);
-        const paidCellValue = paidCell?.value ? paidCell?.value?.toString()?.trim() : 'NO';
-        const paid = paidCellValue === 'YES';
-
-        const paidMonths = MONTHS.reduce((acc, month) => {
-            acc[`${month}_paid`] = paid;
-            return acc;
-        }, {} as Record<string, boolean>);
-
-        const totalColumn = paidColumn + 1;
-        const totalCell = workSheet.cell(row, totalColumn);
-        const totalCellValue = totalCell?.value ? totalCell?.value?.toString()?.trim() : '0';
-        const total = Number(totalCellValue) || 0;
-
-        const monthValuesObj = Object.assign({}, ...monthValues);
-
-        const bodyData = {
-            ...monthsObj,
-            ...monthValuesObj,
-            ...paidMonths,
-            year: bill.year,
-            bill,
-            name,
-            supplier,
-            is_aggregate: !isParent,
-            aggregate_name: cleanTextByListText([bill.name, supplier], name) ?? '',
-            paid,
-            total
-        };
-
-        return { data: bodyData, supplierList: normalizedSupplierList };
-    }
-
-    public calculateAll(expenses: Array<Expense> = []): AllExpensesCalculated {
-        const rawTotal = expenses.reduce((acc, expense) => {
-            const total = expense?.months?.reduce((acc, item) => acc + item.value, 0) || 0;
-            return acc + total;
-        }, 0);
-        const allPaid = expenses.every((expense) => expense.months.every((month) => month.paid));
-        const rawTotalPaid = expenses.reduce((acc, expense) => {
-            const total = expense?.months?.reduce((acc, item) => acc + (item.paid ? item.value : 0), 0) || 0;
-            return acc + total;
-        }, 0);
-        const total = Number(rawTotal.toFixed(2));
-        const totalPaid = Number(rawTotalPaid.toFixed(2));
-        const rawTotalPending = total - totalPaid;
-        const totalPending = Number(rawTotalPending.toFixed(2))
-        return { total, allPaid, totalPaid, totalPending };
+        }, {
+            total: 0,
+            allPaid: true,
+            totalPaid: 0,
+            totalPending: 0,
+        } as MonthsCalculated);
     }
 
     public convertMonthsToObject(expense: Expense): ExpenseWithMonthsAndPaid {
-        const initialObjectMonths: ExpenseMonthsWithPaid = {
-            january: 0,
-            january_paid: false,
-            february: 0,
-            february_paid: false,
-            march: 0,
-            march_paid: false,
-            april: 0,
-            april_paid: false,
-            may: 0,
-            may_paid: false,
-            june: 0,
-            june_paid: false,
-            july: 0,
-            july_paid: false,
-            august: 0,
-            august_paid: false,
-            september: 0,
-            september_paid: false,
-            october: 0,
-            october_paid: false,
-            november: 0,
-            november_paid: false,
-            december: 0,
-            december_paid: false,
-        };
-        const objectMonths = expense?.months?.reduce((acc, month) => {
-            const currentMonth = month.label.toLowerCase() as TMonth;
-            acc[currentMonth] = month.value;
-            acc[`${currentMonth}_paid`] = month.paid;
-            return acc;
-        }, initialObjectMonths);
+        const objectMonths = this.monthsBusiness.convertMonthsToObject(expense.months)
 
         const result: ExpenseWithMonthsAndPaid = {
             ...objectMonths,
@@ -605,11 +168,11 @@ export default class ExpenseBusiness {
             children: undefined,
         }
 
-        if(expense.parent) {
+        if (expense.parent) {
             result.parent = this.convertMonthsToObject(expense.parent);
         }
 
-        if(expense.children && expense.children.length) {
+        if (expense.children && expense.children.length) {
             result.children = expense.children.map((child) => this.convertMonthsToObject(child));
         }
