+++ packages/business/src/finance/expense/business/spreadsheet/index.ts
@@ -1,0 +1,2 @@
+export { default as SpreadsheetBusiness } from './spreadsheet';
+export * from './types';
