+++ packages/business/src/finance/expense/service/service.spec.ts
@@ -88,6 +88,7 @@ describe('Expense Service', () => {
                         create: jest.fn(),
                         update: jest.fn(),
                         delete: jest.fn(),
+                        getAllByBill: jest.fn(),
                     },
                 },
             },
@@ -186,7 +187,7 @@ describe('Expense Service', () => {
     });
 
     describe('getAll', () => {
-        it('should successfully getAll supplier list', async () => {
+        it('should successfully getAll expense list', async () => {
             mockNest.finance.bill.expense.getAll.mockResolvedValue(
                 mockEntityList,
             );
@@ -199,7 +200,7 @@ describe('Expense Service', () => {
             expect(result).toEqual(mockEntityList);
         });
 
-        it('should successfully getAll supplier list paginate', async () => {
+        it('should successfully getAll expense list paginate', async () => {
             mockNest.finance.bill.expense.getAll.mockResolvedValue(
                 mockEntityPaginate,
             );
@@ -211,6 +212,38 @@ describe('Expense Service', () => {
             expect(mockNest.finance.bill.expense.getAll).toHaveBeenCalledWith(
                 mockPaginateParams,
                 mockEntity.bill.id,
+            );
+            expect(result).toEqual(mockEntityPaginate);
+        });
+    });
+
+    describe('getAllByBill', () => {
+        it('should successfully getAll expense list by bill id', async () => {
+            mockNest.finance.bill.expense.getAllByBill.mockResolvedValue(
+                mockEntityList,
+            );
+            const result = await service.getAllByBill(mockEntity.bill.id, {});
+
+            expect(mockNest.finance.bill.expense.getAllByBill).toHaveBeenCalledWith(
+                mockEntity.bill.id,
+                {},
+            );
+            expect(result).toEqual(mockEntityList);
+        });
+
+        it('should successfully getAll expense list paginate by bill id', async () => {
+            mockNest.finance.bill.expense.getAllByBill.mockResolvedValue(
+                mockEntityPaginate,
+            );
+            const result = await service.getAllByBill(
+                mockEntity.bill.id,
+                mockPaginateParams,
+
+            );
+
+            expect(mockNest.finance.bill.expense.getAllByBill).toHaveBeenCalledWith(
+                mockEntity.bill.id,
+                mockPaginateParams,
             );
             expect(result).toEqual(mockEntityPaginate);
         });
